# generic_server: throttle and shed incoming connections according to semaphore limit