--- a/generic_server.cc
+++ b/generic_server.cc
@@ -129,6 +129,10 @@ future<> connection::process()
     });
 }
 
+void connection::on_connection_ready()
+{
+}
+
 void connection::on_connection_close()
 {
 }