--- a/generic_server.hh
+++ b/generic_server.hh
@@ -12,7 +12,9 @@
 #include "utils/log.hh"
 
 #include "seastarx.hh"
+#include "utils/updateable_value.hh"
 
+#include <cstdint>
 #include <list>
 
 #include <seastar/core/file-types.hh>
@@ -41,6 +43,12 @@ public:
     using connection_process_loop = noncopyable_function<future<> ()>;
     using execute_under_tenant_type = noncopyable_function<future<> (connection_process_loop)>;
     bool _tenant_switch = false;
+    struct cpu_concurrency_t {
+        named_semaphore& semaphore;
+        semaphore_units<named_semaphore_exception_factory> units;
+        bool stopped;
+    };
+    cpu_concurrency_t _conns_cpu_concurrency;
     execute_under_tenant_type _execute_under_current_tenant = no_tenant();
 protected:
     server& _server;
@@ -54,7 +62,7 @@ protected:
 private:
     future<> process_until_tenant_switch();
 public:
-    connection(server& server, connected_socket&& fd);
+    connection(server& server, connected_socket&& fd, named_semaphore& sem, semaphore_units<named_semaphore_exception_factory> initial_sem_units);
     virtual ~connection();
 
     virtual future<> process();
@@ -115,7 +123,10 @@ protected:
     std::list<gentle_iterator> _gentle_iterators;
     std::vector<server_socket> _listeners;
     shared_ptr<seastar::tls::server_credentials> _credentials;
-
+private:
+    utils::updateable_value<uint32_t> _conns_cpu_concurrency;
+    uint32_t _prev_conns_cpu_concurrency;
+    named_semaphore _conns_cpu_concurrency_semaphore;
 public:
     server(const sstring& server_name, logging::logger& logger, config cfg);
 
@@ -140,7 +151,7 @@ public:
     future<> do_accepts(int which, bool keepalive, socket_address server_addr);
 
 protected:
-    virtual seastar::shared_ptr<connection> make_connection(socket_address server_addr, connected_socket&& fd, socket_address addr) = 0;
+    virtual seastar::shared_ptr<connection> make_connection(socket_address server_addr, connected_socket&& fd, socket_address addr, named_semaphore& sem, semaphore_units<named_semaphore_exception_factory> initial_sem_units) = 0;
 
     virtual future<> advertise_new_connection(shared_ptr<connection> conn);