--- a/generic_server.hh
+++ b/generic_server.hh
@@ -74,8 +74,6 @@ public:
 
     virtual void on_connection_ready();
 
-    virtual void on_connection_close();
-
     virtual future<> shutdown();
 
     void switch_tenant(execute_under_tenant_type execute);