--- a/transport/server.cc
+++ b/transport/server.cc
@@ -246,7 +246,7 @@ cql_server::cql_server(distributed<cql3::query_processor>& qp, auth::service& au
         service::memory_limiter& ml, cql_server_config config, const db::config& db_cfg,
         qos::service_level_controller& sl_controller, gms::gossiper& g, scheduling_group_key stats_key,
         maintenance_socket_enabled used_by_maintenance_socket)
-    : server("CQLServer", clogger)
+    : server("CQLServer", clogger, generic_server::config(db_cfg))
     , _query_processor(qp)
     , _config(std::move(config))
     , _max_request_size(_config.max_request_size)