@kostja 2025-02-13
Same comment applies to the counting itself. It's best to bump the connection counter as early as possible.
The counter itself should be a RAII, otherwise it may easily leak on exception/destruction.