@gleb-cloudius 2025-02-13
Is there an issue the patch fixes? How drivers handle those connection drops?