@mykaul 2025-02-13
> While this is certainly better than nothing, my concern is that at the time we process the STARTUP message it might already be too late. Can we shed earlier? Basically as early as possible in the connection life cycle.

STARTUP is quite early, it's before authentication I believe and just after OPTIONS/SUPPORTED (which I thought we should cache - https://github.com/scylladb/scylladb/issues/14386 )