@nuivall 2025-02-13
v2:
- fixed documentation
- moved shedding lower in the stack (generic_server)
- RAII counting (gate)