@nuivall 2025-02-13
> Is there an issue the patch fixes? 

Issue added, there is also enterprise issue https://github.com/scylladb/scylla-enterprise/issues/5210 (not adding the link in commit message as it's OS patch). 

> How drivers handle those connection drops?

They will retry, as they already do (hence the storm)