@gleb-cloudius 2025-02-13
> > Is there an issue the patch fixes?
> 
> Issue added, there is also enterprise issue [scylladb/scylla-enterprise#5210](https://github.com/scylladb/scylla-enterprise/issues/5210) (not adding the link in commit message as it's OS patch).

Is this correct issue? This one is about security.

> 
> > How drivers handle those connection drops?
> 
> They will retry, as they already do (hence the storm)

So how dropping connections help if it does not stop the storm?