@nuivall 2025-02-13
> > > Is there an issue the patch fixes?
> > 
> > 
> > Issue added, there is also enterprise issue [scylladb/scylla-enterprise#5210](https://github.com/scylladb/scylla-enterprise/issues/5210) (not adding the link in commit message as it's OS patch).
> 
> Is this correct issue? This one is about security.
> 

Sorry wrong one, corrected.

> > > How drivers handle those connection drops?
> > 
> > 
> > They will retry, as they already do (hence the storm)
> 
> So how dropping connections help if it does not stop the storm?

- limit work per connection, currently we only send at the end of processing, so if client aborts we will do all the work and then receive broken pipe exception
- what we want here is to increase breaking point
- we had to rely on driver's backoff to stop the storm
- we'd like to make progress with some connections
- with thousands of driver instances we need to survive the initial blow, in theory even if we don't make progress, but just avoid clogging task queue, drivers will start to increase retry interval