@gleb-cloudius 2025-02-13
> > > > How drivers handle those connection drops?
> > > 
> > > 
> > > They will retry, as they already do (hence the storm)
> > 
> > 
> > So how dropping connections help if it does not stop the storm?

It looks like I am not the only one having this concern https://github.com/scylladb/scylla-enterprise/issues/5185#issuecomment-2656405926

> 
>     * decrease cpu work per new connection

Ok. But eventually the work will have to be done since a client will retry.

> 
>     * what we want here is to increase breaking point
> 
>     * we have to rely on driver's backoff to stop the storm
> 

Do we know how this backoff works? May be we need to have a backoff in the shedding as well (sleep before dropping a connection).

>     * we'd like to make progress with some connections
> 
>     * with thousands of driver instances we need to survive the initial blow, in theory even if we don't make progress, but just avoid clogging task queue, drivers will start to increase retry interval

Do we know if connection attempt is ever timeouts on a driver side? If yes after how much time? If no limiting authentication parallelism should fix the issue, no?